fix mem_utils_calloc stub signature to match SDK's size_t param
The SDK's app_mem_utils.h declares mem_utils_calloc with a size_t size
parameter, but the unit test stub defined it with uint16_t, which GCC
rejects as a conflicting redeclaration when building against the
current SDK headers.

+++ tests/unit/mock_sources/mem_utils_stub.c
@@ -66,7 +66,7 @@ void mem_utils_free_and_null(void **buffer, const char *file, int line) {
     }
 }
 
-bool mem_utils_calloc(void **buffer, uint16_t size, bool permanent, const char *file, int line) {
+bool mem_utils_calloc(void **buffer, size_t size, bool permanent, const char *file, int line) {
     (void) permanent;
     (void) file;
     (void) line;
